test(oauth): add comprehensive unit tests for OAuth functionality, including access token validation, audience resolution, and rate limiting
